Bootstrap application via bootstrapApplication
Other:
 - Add standalone as default config in schematics
 - Separate transloco loader as its own file
 - Add shared index file
 - Delete shared module
