+++ src/app/app.component.ts
@@ -1,7 +1,11 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
+import { NavigationComponent } from '@app/navigation/navigation.component';
 
 @Component({
   selector: 'app-root',
+  standalone: true,
+  imports: [CommonModule, NavigationComponent],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
 })
