+++ src/app/core/translation/transloco-http-loader.ts
@@ -1,0 +1,15 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { Translation, TranslocoLoader } from '@ngneat/transloco';
+import { Observable } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class TranslocoHttpLoader implements TranslocoLoader {
+  constructor(private http: HttpClient) {}
+
+  public getTranslation(lang: string): Observable<Translation> {
+    return this.http.get<Translation>(`/assets/i18n/${lang}.json`);
+  }
+}
