+++ src/app/navigation/menu/navigation-menu.component.ts
@@ -1,16 +1,24 @@
+import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
 import { RouterModule } from '@angular/router';
 import { RouteItems } from '@app/core/constants/route.constants';
-import { SharedModule } from '@app/shared/shared.module';
+import { TranslocoModule } from '@ngneat/transloco';
 import packageInfo from 'package.json';
 
 @Component({
   selector: 'app-navigation-menu',
   standalone: true,
-  imports: [MatButtonModule, MatIconModule, MatMenuModule, RouterModule, SharedModule],
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatIconModule,
+    MatMenuModule,
+    RouterModule,
+    TranslocoModule,
+  ],
   templateUrl: './navigation-menu.component.html',
   styleUrls: ['./navigation-menu.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
