+++ src/app/routine/runner/timer/runner-timer.component.ts
@@ -1,11 +1,11 @@
+import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, HostBinding, Input, OnInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { Routine, RoutineStep } from '@app/core/models/routine';
 import { Runner } from '@app/core/models/runner';
-import { SecondsToTimePipe } from '@app/shared/pipes/seconds-to-time.pipe';
-import { SharedModule } from '@app/shared/shared.module';
+import { SecondsToTimePipe } from '@app/shared';
 import { RxState } from '@rx-angular/state';
 import { PushModule } from '@rx-angular/template/push';
 import { map, Observable, of, switchMap } from 'rxjs';
@@ -20,12 +20,12 @@ export interface RunnerTimerComponentState {
   selector: 'app-runner-timer',
   standalone: true,
   imports: [
+    CommonModule,
     MatButtonModule,
     MatIconModule,
     MatProgressSpinnerModule,
     PushModule,
     SecondsToTimePipe,
-    SharedModule,
   ],
   templateUrl: './runner-timer.component.html',
   styleUrls: ['./runner-timer.component.scss'],
