+++ src/app/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './directives/back-button.directive';
+export * from './pipes/seconds-to-time.pipe';
+export * from './utils/string-utils';
