--- src/app/shared/shared.module.ts
@@ -1,9 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { TranslocoModule } from '@ngneat/transloco';
-
-@NgModule({
-  imports: [CommonModule, TranslocoModule],
-  exports: [CommonModule, TranslocoModule],
-})
-export class SharedModule {}
