--- src/app/transloco-root.module.ts
@@ -1,37 +1,0 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable, isDevMode, NgModule } from '@angular/core';
-import {
-  Translation,
-  TRANSLOCO_CONFIG,
-  TRANSLOCO_LOADER,
-  translocoConfig,
-  TranslocoLoader,
-  TranslocoModule,
-} from '@ngneat/transloco';
-import { Observable } from 'rxjs';
-
-@Injectable({ providedIn: 'root' })
-export class TranslocoHttpLoader implements TranslocoLoader {
-  constructor(private http: HttpClient) {}
-
-  public getTranslation(lang: string): Observable<Translation> {
-    return this.http.get<Translation>(`/assets/i18n/${lang}.json`);
-  }
-}
-
-@NgModule({
-  exports: [TranslocoModule],
-  providers: [
-    {
-      provide: TRANSLOCO_CONFIG,
-      useValue: translocoConfig({
-        availableLangs: ['en', 'es'],
-        defaultLang: 'en',
-        reRenderOnLangChange: true,
-        prodMode: !isDevMode(),
-      }),
-    },
-    { provide: TRANSLOCO_LOADER, useClass: TranslocoHttpLoader },
-  ],
-})
-export class TranslocoRootModule {}
